# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "会議室管理システムのAPI",
        "title": "Elmo Project API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/participants": {
            "get": {
                "description": "指定された会議室の参加者一覧を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participants"
                ],
                "summary": "参加者一覧を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会議室ID",
                        "name": "room_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ParticipantsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "指定された会議室に参加者を追加します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "participants"
                ],
                "summary": "参加者を追加",
                "parameters": [
                    {
                        "description": "参加者情報",
                        "name": "participant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParticipantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ParticipantRequest": {
            "type": "object",
            "properties": {
                "room_id": {
                    "type": "string",
                    "example": "room123"
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "models.ParticipantUser": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "user123"
                },
                "name": {
                    "type": "string",
                    "example": "田中太郎"
                }
            }
        },
        "models.ParticipantsResponse": {
            "type": "object",
            "properties": {
                "room_id": {
                    "type": "string",
                    "example": "room123"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ParticipantUser"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}